优化节点选择效率 预埋服务统计

+++ src/main/java/org/song/qsrpc/receiver/NodeRegistry.java
@@ -40,7 +40,7 @@ public class NodeRegistry {
         nodeInfo.setIp(node_ip);
         nodeInfo.setPort(port);
         if (weight != null) {
-            nodeInfo.setWeight(Integer.parseInt(weight));
+            nodeInfo.setWeight(Byte.parseByte(weight));
         }
         nodeInfo.setZkIps(zkIps);
         nodeInfo.setZkPath(zkPath);
